change name for controller and entity, to match the animal them. Also added a index html with me change a lot of names to match the them
