+++ src/main/resources/templates/Index.html
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://thymeleaf.org">
+<head>
+<meta charset="ISO-8859-1">
+<title>Animal Index</title>
+</head>
+<body>
+<div class="container">
+
+		<h4>Animal Index</h4>
+		<hr>
+		<table class="table table-bordered">
+			<thead>
+				<tr class="bg-danger">
+					<th>Animal</th>
+					<th>kingdom</th>
+				</tr>
+			</thead>
+	</div>
+</body>
+</html>
